reconciled with configure_fingerprint now using self-signed cert for dtls browsers dont seem to care

+++ boringssl/ssl/test/async_bio.cc
@@ -102,7 +102,7 @@ static long AsyncCtrl(BIO *bio, int cmd, long num, void *ptr) {
     return 0;
   }
   BIO_clear_retry_flags(bio);
-  long ret = BIO_ctrl(bio->next_bio, cmd, num, ptr);
+  int ret = BIO_ctrl(bio->next_bio, cmd, num, ptr);
   BIO_copy_next_retry(bio);
   return ret;
 }
